[init] switch from packer.nvim to lazy.nvim

+++ init.lua
@@ -1,21 +1,34 @@
-local packer_path = vim.fn.stdpath("data") .. "/site/pack"
-function ensure (user, repo)
-  local install_path = string.format("%s/packer/start/%s", packer_path, repo, repo)
-  if vim.fn.empty(vim.fn.glob(install_path)) > 0 then
-    vim.api.nvim_command(string.format("!git clone https://github.com/%s/%s %s", user, repo, install_path))
-    vim.api.nvim_command(string.format("packadd %s", repo))
+local lazy_path = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+local package_path = vim.fn.stdpath("data") .. "/lazy"
+if not vim.loop.fs_stat(lazy_path) then
+  vim.fn.system({
+    "git",
+    "clone",
+    "--filter=blob:none",
+    "--single-branch",
+    "https://github.com/folke/lazy.nvim.git",
+    lazy_path,
+  })
+end
+vim.opt.runtimepath:prepend(lazy_path)
+function ensure (repo, package, dir)
+  if not dir then
+    vim.fn.system({
+      "git",
+      "clone",
+      "--filter=blob:none",
+      "--single-branch",
+      repo,
+      package_path,
+    })
+    vim.api.nvim_command(string.format("packadd %s", package))
+  else
+    local install_path = string.format("%s/%s", package_path, package)
+    vim.fn.system(string.format("rm -r %s", install_path))
+    vim.fn.system(string.format("ln -s %s %s", repo, package_path))
+    vim.api.nvim_command(string.format("packadd %s", package))
   end
 end
-function local_ensure (dir, repo)
-    local install_path = string.format("%s/packer/start/%s", packer_path, repo)
-    local pack_path = string.format("%s/packer/start/", packer_path)
-    vim.fn.system(string.format("rm -r %s", install_path))
-    vim.fn.system(string.format("ln -s %s %s", dir, pack_path))
-    vim.api.nvim_command(string.format("packadd %s", repo))
-end
-ensure("lewis6991", "impatient.nvim")
-require("impatient")
-ensure("wbthomason", "packer.nvim")
-local_ensure("~/Git\\ Repos/katcros-fnl/", "katcros-fnl")
-ensure("Olical", "aniseed")
+ensure("~/Git\\ Repos/katcros-fnl", "katcros-fnl", true)
+ensure("Olical/aniseed", "aniseed")
 vim.g["aniseed#env"] = {module = "init", compile = true}
